+++ src/WelcomingSection/MovieSection.scss
